# Fix identityHashCode for JS and WasmJS