--- a/src/commonMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.kt
+++ b/src/commonMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.kt
@@ -13,4 +13,39 @@ internal expect fun getEnvironmentVariable(key: String): String?
 @JvmInline
 internal value class IdentityHashCode(private val value: Int)
 
-internal expect fun identityHashCode(any: Any?): IdentityHashCode
+/**
+ * Returns the [IdentityHashCode] for an object. For `null` always returns `IdentityHashCode(0)`
+ * @throws IllegalArgumentException if [any] is a primitive type or a [String]
+ */
+internal fun identityHashCode(any: Any?): IdentityHashCode =
+    when {
+        any == null              -> IdentityHashCode(0)
+        hasIdentityHashCode(any) -> objectIdentityHashCode(any)
+        else                     ->
+            throw IllegalArgumentException("identity hash code cannot be computed for primitives and strings (type: ${any::class.simpleName})")
+    }
+
+internal expect fun objectIdentityHashCode(any: Any): IdentityHashCode
+
+/**
+ * Function returns `true` if the object can be used to compute the [IdentityHashCode] on all platforms.
+ * The following types do not have identity:
+ * + primitive numbers
+ * + boolean
+ * + character
+ * + string (despite the fact that it is an object on most of the platforms but JS does not allow to apply identity hash code logic to it)
+ */
+internal fun hasIdentityHashCode(any: Any?): Boolean =
+    when (any) {
+        is Byte,
+        is Short,
+        is Int,
+        is Long,
+        is Float,
+        is Double,
+        is Boolean,
+        is Char,
+        is String -> false
+
+        else      -> true
+    }