--- a/src/commonTest/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/SystemTest.kt
+++ b/src/commonTest/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/SystemTest.kt
@@ -1,5 +1,6 @@
 package it.krzeminski.snakeyaml.engine.kmp.internal
 
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
@@ -35,5 +36,26 @@ class SystemTest : FunSpec({
         val hashCode1 = identityHashCode(null)
         val hashCode2 = identityHashCode(null)
         hashCode1 shouldBe hashCode2
+        hashCode1 shouldBe IdentityHashCode(0)
+    }
+
+    listOf(
+        1.toByte(),
+        1.toShort(),
+        1.toInt(),
+        1.toLong(),
+        1.5.toFloat(),
+        1.5.toDouble(),
+        "test",
+        'a',
+    ).forEach { primitive ->
+        val type = primitive::class.simpleName
+        test("identityHashCode -> primitive type $type causes exception") {
+            val ex = shouldThrow<IllegalArgumentException> {
+                identityHashCode(primitive)
+            }
+            ex.message shouldBe
+                "identity hash code cannot be computed for primitives and strings (type: $type)"
+        }
     }
 })