--- a/src/commonTest/kotlin/it/krzeminski/snakeyaml/engine/kmp/representer/IdentityHashCodeEnabledOrReasonIf.kt
+++ b/src/commonTest/kotlin/it/krzeminski/snakeyaml/engine/kmp/representer/IdentityHashCodeEnabledOrReasonIf.kt
@@ -7,10 +7,9 @@ import io.kotest.core.test.EnabledOrReasonIf
 
 internal val identityHashCodeEnabledOrReasonIf: EnabledOrReasonIf = {
     when (platform) {
-        Platform.JS,
-        Platform.WasmJs ->
+        Platform.JS ->
             Enabled.disabled("identity hashcode does not work correctly: https://github.com/krzema12/snakeyaml-engine-kmp/pull/273")
 
-        else            -> Enabled.enabled
+        else        -> Enabled.enabled
     }
 }