--- a/src/jsMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.js.kt
+++ b/src/jsMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.js.kt
@@ -7,12 +7,13 @@ package it.krzeminski.snakeyaml.engine.kmp.internal
 @JsName("Symbol")
 private external fun symbol(name: String): dynamic
 
-private var lastIdentityHashCodeId = 0
+// zero is reserved for null
+private var lastIdentityHashCodeId = 1
 private val IDENTITY_HASH_CODE_SYMBOL = symbol("KotlinIdentityHashCode")
 
 internal actual fun identityHashCode(any: Any?): IdentityHashCode {
-    require(jsTypeOf(any) == "object") { "Only non-primitive types are supported!" }
     if (any == null) return IdentityHashCode(0)
+    require(jsTypeOf(any) == "object") { "Only non-primitive types are supported!" }
     val dyn = any.asDynamic()
     if (dyn[IDENTITY_HASH_CODE_SYMBOL] === undefined) {
         dyn[IDENTITY_HASH_CODE_SYMBOL] = lastIdentityHashCodeId++