--- a/src/jsMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.js.kt
+++ b/src/jsMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.js.kt
@@ -11,9 +11,7 @@ private external fun symbol(name: String): dynamic
 private var lastIdentityHashCodeId = 1
 private val IDENTITY_HASH_CODE_SYMBOL = symbol("KotlinIdentityHashCode")
 
-internal actual fun identityHashCode(any: Any?): IdentityHashCode {
-    if (any == null) return IdentityHashCode(0)
-    require(jsTypeOf(any) == "object") { "Only non-primitive types are supported!" }
+internal actual fun objectIdentityHashCode(any: Any): IdentityHashCode {
     val dyn = any.asDynamic()
     if (dyn[IDENTITY_HASH_CODE_SYMBOL] === undefined) {
         dyn[IDENTITY_HASH_CODE_SYMBOL] = lastIdentityHashCodeId++