--- a/src/jvmMain/java/it/krzeminski/snakeyaml/engine/kmp/internal/system.jvm.kt
+++ b/src/jvmMain/java/it/krzeminski/snakeyaml/engine/kmp/internal/system.jvm.kt
@@ -5,7 +5,7 @@ package it.krzeminski.snakeyaml.engine.kmp.internal
  */
 internal actual fun getEnvironmentVariable(key: String): String? = System.getenv(key)
 
-internal actual fun identityHashCode(any: Any?): IdentityHashCode {
+internal actual fun objectIdentityHashCode(any: Any): IdentityHashCode {
     val hc = System.identityHashCode(any)
     return IdentityHashCode(hc)
 }