--- a/src/wasmJsMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.wasm.kt
+++ b/src/wasmJsMain/kotlin/it/krzeminski/snakeyaml/engine/kmp/internal/system.wasm.kt
@@ -5,16 +5,19 @@ internal actual fun identityHashCode(any: Any?): IdentityHashCode {
     if (any == null) return IdentityHashCode(0)
     val ref = any.toJsReference()
     if (!identityHashCodes.has(ref)) {
-        identityHashCodes[ref] = lastIdentityHashCodeId++
+        identityHashCodes.set(ref, lastIdentityHashCodeId++)
     }
-    val hc = identityHashCodes[ref]
+    val hc = identityHashCodes.get(ref)
     return IdentityHashCode(hc)
 }
 
-private external interface IdentityHashCodeMap {
-    operator fun get(key: JsReference<Any>): Int
-    operator fun set(key: JsReference<Any>, value: Int)
-    fun has(key: JsReference<Any>): Boolean
+// NOTE: don't add operator modifier to get and set methods
+// kotlin transform those method into `map[key]`
+// instead of `map.get(key)` and causes exception to be thrown
+private external class IdentityHashCodeMap : JsAny {
+    fun get(key: JsAny): Int
+    fun set(key: JsAny, value: Int)
+    fun has(key: JsAny): Boolean
 }
 
 private val identityHashCodes: IdentityHashCodeMap = js("new WeakMap()")