Part of https://github.com/krzema12/snakeyaml-engine-kmp/issues/279

TODO:
* [x] Add tests that cover this piece for all targets