@OptimumCode 2024-12-27
Hi, @krzema12 @aSemy. What do you think about implementing identity hash code using WeakMap for JS platform if the parameter is not an `object` (similar to wasmJs)? Right now, I think, this is the only thing that blocks this PR from going forward