@krzema12 2024-12-27
Yeah, worth trying, I'm just focusing on porting other simpler changes for now. If you want to tackle it, feel welcome!