@OptimumCode 2024-12-27
Turns out, you can only use objects or non-registered symbols as a key for WeakMap (https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WeakMap/set#key). Looking for another way to make it work