@krzema12 2024-12-27
@OptimumCode so the problematic piece is the primitive types, right? Have you considered converting them to something that can be used as a WeakMap key so that the 1:1 relation is preserved?