@OptimumCode 2024-12-27
> Have you considered converting them to something that can be used as a WeakMap key so that the 1:1 relation is preserved?

Yes, but honestly I don't know what could be used here since creating an object would mean that each primitive value would have a unique identity. We can simply increment the internal counter we use on every primitive type instead:

```kotlin
    if (jsTypeOf(any) != "object") {
        return IdentityHashCode(lastIdentityHashCodeId++)
    }
```

I was also thinking about an alternative approach:

The main reason for having the identity map in the representer is to prevent recursion. Right now, the implementation uses `IdentityHashCode` class as a key (value type wrapping `Int`). Since the primitive types and strings cannot cause the recursion in any case, we could change what we use as a key in the map. If the key is a primitive or a string - we use the object itself as a key. Otherwise, we create a wrapper object containing an identity hashcode for the key and store this object instead. Something like this:

```kotlin

val map: MutableMap<Any, Node> = ...
val key: Any? = ...
if (isPrimitiveOrString(key)) {
  map[key] = ...
} else {
  map[IdentityHashCode(identityHashCode(key))] = ...
}
```