@krzema12 2024-12-27
> Yes, but honestly I don't know what could be used here since creating an object would mean that each primitive value would have a unique identity.

I was thinking about dynamically building a symbol name, with gluing the primitive's type and value. I don't know how much sense it makes in practice, though.

Your idea sounds good! Definitely worth testing.