@OptimumCode 2024-12-28
Considering the error of measurement it look more or less okay
```
Before:

macosX64.DumpBenchmark.dumpMapWithAllTypes ( {} )	91.7295171144458 us/op		
mingwX64.DumpBenchmark.dumpMapWithAllTypes ( {} )	85.99261374108332 us/op		
macosArm64.DumpBenchmark.dumpMapWithAllTypes ( {} )	21.850650706950642 us/op		
jvm.DumpBenchmark.dumpMapWithAllTypes	                15.243241587097435 us/op		
js.DumpBenchmark.dumpMapWithAllTypes ( {} )            115.11791006436947 us/op		
linuxX64.DumpBenchmark.dumpMapWithAllTypes ( {} )	52.52259395065473 us/op		

After:

macosX64.DumpBenchmark.dumpMapWithAllTypes ( {} )	67.95420038861793 us/op		
mingwX64.DumpBenchmark.dumpMapWithAllTypes ( {} )	86.26141539258876 us/op		
macosArm64.DumpBenchmark.dumpMapWithAllTypes ( {} )	20.78312291351127 us/op		
jvm.DumpBenchmark.dumpMapWithAllTypes	                17.498812364098967 us/op		
js.DumpBenchmark.dumpMapWithAllTypes ( {} )             98.66989433118133 us/op		
linuxX64.DumpBenchmark.dumpMapWithAllTypes ( {} )	56.87801296996396 us/op		
```